requirements.txt
Flask
tensorflow
keras
Pillow
pandas
numpy
